+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `Stickers` test cases
+- `StickerOwnerUserId` parameter to Systems Integration Tests configs
+- Exception `InvalidStickerSetNameException`
+- Exception `InvalidStickerEmojisException`
+- Exception `InvalidStickerDimensionsException`
+- Exception `StickerSetNameExistsException`
+
 ## [13.4.0-rc2] - 2017-11-26
 
 ### Fixed